// car-gal/interfaces.ts
export interface IcGElementStyleObject {
    id?: string | string[];
    values?: string[][];
    childValues?: IcGElementStyleObject[];
}

export interface ICgElement {
    //get ele?: HTMLElement;
    element?: HTMLElement;
    parentElement?: HTMLElement;
    tagName?: string;
    textContent?: string;
    classes?: string;
    //styles?: string[][];
    styles?: IcGElementStyleObject;
    children?: (CgElement | ICgElement | undefined)[];
    eventListeners?: InyGalleryEventListener[];
    attr?: string[][];
    removeOnDispose?: boolean;
}

export interface InyGalleryEventListener {
    action: string;
    handler: (event: UIEvent) => void;
}